+++ src/main/java/net/additionz/access/PassiveAgeAccess.java
@@ -2,6 +2,6 @@ package net.additionz.access;
 
 public interface PassiveAgeAccess {
 
-    public int getPassiveAge();
+    boolean isTeenager();
 
 }