+++ src/main/java/net/additionz/waila/AgeJadeProvider.java
@@ -2,7 +2,6 @@ package net.additionz.waila;
 
 import net.additionz.AdditionMain;
 import net.additionz.access.PassiveAgeAccess;
-import net.minecraft.entity.passive.PassiveEntity;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 import snownee.jade.api.EntityAccessor;
@@ -20,8 +19,9 @@ public enum AgeJadeProvider implements IEntityComponentProvider {
 
     @Override
     public void appendTooltip(ITooltip tooltip, EntityAccessor accessor, IPluginConfig config) {
-        int realPassiveAge = (int) Math.round(Math.floor(((PassiveAgeAccess) accessor.getEntity()).getPassiveAge() / AdditionMain.CONFIG.passiveEntityConfig.passive_age_calculation)) + 1;
-        tooltip.append(Text.translatable("entity.additionz.age_jade.tooltip", ((PassiveEntity) accessor.getEntity()).isBaby() ? 0 : realPassiveAge));
+        if (((PassiveAgeAccess) accessor.getEntity()).isTeenager()) {
+            tooltip.append(Text.translatable("entity.additionz.age_jade.tooltip"));
+        }
     }
 
 }